update op squeeze
