+++ cinn/hlir/op/elementwise.cc
@@ -514,6 +514,73 @@ StrategyForUnary(sigmoid, Sigmoid);
 
 #undef StrategyForUnary
 
+std::shared_ptr<framework::OpStrategy> StrategyForSqueeze(const framework::NodeAttr &attrs,
+                                                          const std::vector<ir::Tensor> &inputs,
+                                                          const std::vector<Type> &out_type,
+                                                          const std::vector<std::vector<int>> &output_shapes,
+                                                          const Target &target) {
+  CHECK(attrs.attr_store.count("axes")) << "find no attr of axes";
+  std::vector<int> axes = absl::get<std::vector<int>>(attrs.attr_store.at("axes"));
+
+  framework::CINNCompute squeeze_compute([=](lang::Args args, lang::RetValue *ret) {
+    CHECK(!args.empty()) << "The input arguments of Squeeze compute is empty! Please check.\n";
+    CINNValuePack pack_args = args[0];
+    CHECK_GE(pack_args.size(), 1U) << "at least 1 input tensors for Squeeze compute\n";
+    Expr A = pack_args[0];
+    CHECK(A.as_tensor());
+    CHECK(!output_shapes.empty());
+    auto tensor_A = A.as_tensor_ref();
+    auto stages   = CreateStages({tensor_A});
+    VLOG(3) << "A shape: " << utils::Join(tensor_A->shape, ", ")
+            << ", output_shapes: " << utils::Join(output_shapes[0], ", ");
+
+    std::string tensor_name = UniqName("Squeeze_out");
+    if (FLAGS_cinn_ir_schedule) {
+      CHECK_EQ(pack_args.size(), 2U);
+      tensor_name = pack_args[1].operator std::string();
+    }
+
+    ir::Tensor out = pe::Squeeze(tensor_A, axes, tensor_name);
+    std::vector<CINNValue> res;
+    stages->InsertLazily(out);
+    res.push_back(CINNValue(out));
+    CHECK(!out_type.empty()) << "Output type of Squeeze is empty! Please check.\n";
+    res.push_back(CINNValue(stages));
+    *ret = CINNValuePack{res};
+  });
+
+  auto strategy = std::make_shared<framework::OpStrategy>();
+  strategy->AddImpl(
+      squeeze_compute, framework::GetInjectiveScheduleFunc(output_shapes, target), "strategy.squeeze.x86", 1);
+  return strategy;
+}
+
+std::vector<std::vector<int>> InferShapeForSqueeze(const std::vector<std::vector<int>> &inputs_shape,
+                                                   const framework::AttrMapType &attrs) {
+  CHECK_EQ(inputs_shape.size(), 1U);
+  std::vector<int> axes = {};  // absl::get<std::vector<int>>(attrs.at("axes"));
+
+  std::vector<int> output_shape;
+  if (axes.size()) {
+    for (int idx = 0; idx < inputs_shape[0].size(); ++idx) {
+      // if can't find idx in axis
+      if (std::find(axes.begin(), axes.end(), idx) == axes.end()) {
+        output_shape.push_back(inputs_shape[0][idx]);
+      } else {
+        CHECK_EQ(inputs_shape[0][idx], 1);
+      }
+    }
+  } else {
+    for (int idx = 0; idx < inputs_shape[0].size(); ++idx) {
+      if (inputs_shape[0][idx] != 1) {
+        output_shape.push_back(inputs_shape[0][idx]);
+      }
+    }
+  }
+
+  return {output_shape};
+}
+
 }  // namespace op
 }  // namespace hlir
 }  // namespace cinn
@@ -640,6 +707,17 @@ CINN_REGISTER_HELPER(elementwise_ops) {
 #ifndef CINN_WITH_CUDA
       .set_attr("inferlayout", MakeOpFunction(cinn::hlir::op::InferLayoutForAssignValue))
 #endif
+      .set_attr<cinn::hlir::framework::OpPatternKind>("OpPattern", cinn::hlir::framework::OpPatternKind::kElementWise)
+      .set_support_level(4);
+
+  CINN_REGISTER_OP(squeeze)
+      .describe("The operator is used to squeeze input tensor's dims")
+      .set_num_inputs(1)
+      .set_num_outputs(1)
+      .set_attr<cinn::hlir::framework::StrategyFunction>("CINNStrategy", cinn::hlir::op::StrategyForSqueeze)
+      .set_attr("infershape", MakeOpFunction(cinn::hlir::op::InferShapeForSqueeze))
+      .set_attr("inferdtype", MakeOpFunction(cinn::hlir::op::InferDtypeForElementwise))
+      .set_attr("inferlayout", MakeOpFunction(cinn::hlir::op::InferLayoutForElementwise))
       .set_attr<cinn::hlir::framework::OpPatternKind>("OpPattern", cinn::hlir::framework::OpPatternKind::kElementWise)
       .set_support_level(4);
