+++ cinn/hlir/op/use_ops.h
@@ -26,7 +26,6 @@ CINN_USE_REGISTER(gather_ops)
 CINN_USE_REGISTER(scatter_ops)
 CINN_USE_REGISTER(cast_ops)
 CINN_USE_REGISTER(sort_ops)
-CINN_USE_REGISTER(squeeze_ops)
 CINN_USE_REGISTER(argmin_ops)
 CINN_USE_REGISTER(argmax_ops)
 CINN_USE_REGISTER(reduce_ops)
